Added Metadata to auth/register
